requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
brotli==1.2.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.2
cycler==0.12.1
datasets==2.21.0
ddgs==9.9.1
dill==0.3.5.1
distro==1.9.0
duckduckgo_search==8.1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.7.4
fake-useragent==2.2.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2024.6.1
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
interchange==2021.0.4
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
modelscope==1.27.1
monotonic==1.6
mpmath==1.3.0
multidict==6.6.2
multiprocess==0.70.13
neo4j==5.28.1
networkx==3.4.2
numpy==1.24.3
openai==1.79.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pansi==2024.11.0
pillow==11.1.0
primp==0.15.0
propcache==0.3.2
py2neo==2021.2.4
pyarrow==20.0.0
pydantic==2.11.0
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.3
pypdf==6.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
responses==0.18.0
safetensors==0.5.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
svgwrite==1.4.3
sympy==1.13.3
tiktoken==0.9.0
tokenizers==0.15.2
torch==2.0.1
tqdm==4.67.1
transformers==4.35.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zhipuai==2.1.5.20250106
